templates/api.c:	Fiwx a bug in gliss_run_and_count_inst(), inst were counted on 32 bits, now 64

+++ templates/api.c
@@ -694,7 +694,7 @@ $(end)
  * */
 uint64_t $(proc)_run_and_count_inst($(proc)_sim_t *sim)
 {
-	unsigned int i = 0;
+	uint64_t i = 0;
     uint32_t num_bloc;
     $(proc)_state_t*   state     = sim->state;
     $(proc)_decoder_t* decoder   = sim->decoder;
@@ -870,7 +870,7 @@ $(end)
  * */
 uint64_t $(proc)_run_and_count_inst($(proc)_sim_t *sim)
 {
-	unsigned int i = 0;
+	uint64_t i = 0;
     $(proc)_state_t*   state     = sim->state;
     $(proc)_decoder_t* decoder   = sim->decoder;
     $(proc)_address_t  addr_exit = sim->addr_exit;
